Update globals.py to include "sail_running = False"
I was encountering an error when running a new Sail where it wouldn't save my changes to the sailing configs to settings.dat and was "stuck" using the initially set values. 

From what I could tell this variable was just straight missing from the file? Adding it in and defaulting it to False like this seems to have resolved the issue.